Windows: Avoid dependency on libmingwex.a
While libgcc.a is standalone (and therefore the dependencies on the 64-bit
modulo functions in format7 are fine), libmingwex.a is not. And usleep()
is defined in libmingwex.a. Since the only calls to usleep() on Windows
are in format7.c and avt.c and really want to sleep for microseconds, not
nanoseconds, we can get away by simply redefining usleep() in terms of
Sleep().

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>